[REFACTOR] Remote Entity - Response ~ Header까지는 Base로 분리해도 될 것같음 - Firebase 통신 Entity를 분리 feature/#42
