+++ app/src/main/java/com/sessac/myaitrip/data/entities/remote/TourApiResponse.kt
@@ -3,6 +3,7 @@ package com.sessac.myaitrip.data.entities.remote
 import com.google.gson.annotations.SerializedName
 import com.sessac.myaitrip.data.entities.TourItem
 
+/// BaseResponse
 data class ApiResponse<T>(
     @SerializedName("response")
     var response: Response<T>? = null
@@ -23,6 +24,7 @@ data class Header(
     @SerializedName("resultMsg")
     var resultMsg: String? = null
 )
+/// BaseResponse ( ApiResponse ~ Header )
 
 data class Body<T>(
     @SerializedName("items")
